requirements.txt
tensorflow-cpu==2.8.0
fastapi==0.95.1
opencv-python==4.7.0.72
python-multipart
protobuf==3.20.0
uvicorn
python-dotenv==1.0.0
numpy==1.25.2
pymongo==4.5.0

